refactor(ui): unify transcript widget cards
